* Fixed a leak with spells when doing a reload. * Forgot to lock g_loaderLock. * Warning fix.

+++ otserv/branches/avesta/weapons.cpp
@@ -163,7 +163,7 @@ int32_t Weapons::getMaxMeleeDamage(int32_t attackSkill, int32_t attackValue)
 //players
 int32_t Weapons::getMaxWeaponDamage(int32_t attackSkill, int32_t attackValue, float attackFactor)
 {
-	return std::ceil(((float)(attackSkill * (attackValue * 0.0425) + (attackValue * 0.2)) / attackFactor)) * 2;
+	return ((int32_t)std::ceil(((float)(attackSkill * (attackValue * 0.0425) + (attackValue * 0.2)) / attackFactor)) * 2);
 }
 
 Weapon::Weapon(LuaScriptInterface* _interface) :
